refactor: fix turn-resolution WidgetRef smell and split panel parts (#3878)
Pass narrow GameService/AiProfile/TurnResolutionResultApplier into map
turn-resolution session helpers instead of WidgetRef parameters, and
extract diplomacy negotiation-mood and production allocation row/summary
library parts without changing public behavior.

Refs #3878
