+++ app/lib/features/game/widgets/diplomacy/diplomacy_panel.dart
@@ -35,6 +35,7 @@ part 'diplomacy_panel_chrome_standing.dart';
 part 'diplomacy_panel_constants.dart';
 part 'diplomacy_panel_filter.dart';
 part 'diplomacy_panel_mode_bar.dart';
+part 'diplomacy_panel_order_actions_mood.dart';
 part 'diplomacy_panel_order_actions.dart';
 part 'diplomacy_panel_row.dart';
 part 'diplomacy_panel_row_info.dart';
@@ -74,7 +75,7 @@ class DiplomacyPanel extends StatefulWidget with GamePanelMixin {
 }
 
 class _DiplomacyPanelState extends State<DiplomacyPanel>
-    with _DiplomacyOrderActions {
+    with _DiplomacyOrderActionsMood, _DiplomacyOrderActions {
   final Map<String, String> _moodByLeaderId = <String, String>{};
   final SubscriptionTracker _subscriptions = SubscriptionTracker();
