+++ app/lib/features/game/widgets/diplomacy/diplomacy_panel_order_actions.dart
@@ -3,9 +3,8 @@
 
 part of 'diplomacy_panel.dart';
 
-mixin _DiplomacyOrderActions on State<DiplomacyPanel> {
-  Map<String, String> get moodByLeaderId;
-
+mixin _DiplomacyOrderActions
+    on State<DiplomacyPanel>, _DiplomacyOrderActionsMood {
   void submitOrDialog(DiplomaticOrder order) {
     final pending =
         widget.currentOrders.diplomaticOrdersByPlayerId[widget.humanPlayerId] ??
@@ -125,76 +124,6 @@ mixin _DiplomacyOrderActions on State<DiplomacyPanel> {
       AppendDiplomaticOrderRequestedEvent(
         playerId: widget.humanPlayerId,
         order: order,
-      ),
-    );
-  }
-
-  int pendingCountForTarget(String targetFactionId) {
-    final list =
-        widget.currentOrders.diplomaticOrdersByPlayerId[widget.humanPlayerId] ??
-        const <DiplomaticOrder>[];
-    return list.where((o) => o.targetFactionId == targetFactionId).length;
-  }
-
-  double offerQualityDeltaFor(DiplomaticOrderType type) {
-    switch (type) {
-      case DiplomaticOrderType.declareWar:
-        return -0.8;
-      case DiplomaticOrderType.offerPeace:
-        return 0.6;
-      case DiplomaticOrderType.alliance:
-        return 0.4;
-      case DiplomaticOrderType.breakAlliance:
-        return -0.4;
-      case DiplomaticOrderType.establishOverture:
-        return 0.5;
-      case DiplomaticOrderType.grantAid:
-        return 0.7;
-      case DiplomaticOrderType.setSubsidy:
-        return 0.5;
-      case DiplomaticOrderType.establishFtp:
-        return 0.5;
-      case DiplomaticOrderType.boycott:
-        return -0.6;
-      case DiplomaticOrderType.revokeBoycott:
-        return 0.3;
-    }
-  }
-
-  int stableSeed({
-    required String leaderId,
-    required String discriminator,
-    required int stallCounter,
-  }) {
-    final turn = widget.game.worldState.turnState.turnNumber;
-    final base = widget.game.globalGameSeed ?? 0;
-    final text = '$leaderId|$discriminator|$stallCounter|$turn';
-    var hash = kFnv1aOffsetBasis32;
-    for (final code in text.codeUnits) {
-      hash ^= code;
-      hash = (hash * kFnv1aPrime32) & kDeterministicLcg31Mask;
-    }
-    return base ^ hash;
-  }
-
-  void emitNegotiationMood({
-    required String leaderId,
-    required double offerQualityDelta,
-    required int stallCounter,
-    required String discriminator,
-  }) {
-    final currentMood = moodByLeaderId[leaderId] ?? kDefaultMood;
-    widget.bus.emit(
-      NegotiationMoodUpdateEvent(
-        leaderId: leaderId,
-        currentMood: currentMood,
-        offerQualityDelta: offerQualityDelta,
-        stallCounter: stallCounter,
-        seed: stableSeed(
-          leaderId: leaderId,
-          discriminator: discriminator,
-          stallCounter: stallCounter,
-        ),
       ),
     );
   }
